sửa kiểu dữ liệu của maxspeed

+++ src/com/arkanoid/Const.java
@@ -6,7 +6,7 @@ public final class Const {
 
     public static final int SCREEN_WIDTH = 1000;
     public static final int SCREEN_HEIGHT = 800;
-    public static final int BALL_MAXSPEED = 75;
+    public static final double BALL_MAXSPEED = 75;
     public static final int PADDLE_WIDTH = 120;
     public static final int PADDLE_HEIGHT = 20;
     public static final int BALL_DIAMETER = 20;
